# services/user/sdk/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
        "/api/v1/stock/search": {
            "get": {
                "operationId": "StockController_getSearch",
                "parameters": [
                    {
                        "name": "exchange",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "symbol",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetSearchStockResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
        "/api/v1/stock/{identity}": {
            "get": {
                "operationId": "StockController_get",
                "parameters": [
                    {
                        "name": "identity",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StockDataViewResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
        "/api/v1/stock": {
            "get": {
                "operationId": "StockController_getStocks",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetStockListResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
        "/api/v1/stock/data/csv": {
            "put": {
                "operationId": "StockController_createDataFromCsv",
                "parameters": [],
                "responses": {
                    "409": {
                        "description": "DuplicateEntryError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
        "/api/v1/stock/values/csv": {
            "put": {
                "operationId": "StockController_createValuesFromCsv",
                "parameters": [],
                "responses": {
                    "409": {
                        "description": "DuplicateEntryError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
...
        "/api/v1/user/{userIdentity}/asset/stock": {
            "put": {
                "operationId": "AssetController_createStockAssetForUser",
                "parameters": [
                    {
                        "name": "userIdentity",
                        "required": true,
                        "in": "path",
                        "description": "User identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStockAssetBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateStockAssetForUserResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "DuplicateEntryError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
...
        "/api/v1/assetGroup/{assetGroupIdentity}/asset/stock": {
            "put": {
                "operationId": "AssetController_createStockAssetForGroup",
                "parameters": [
                    {
                        "name": "assetGroupIdentity",
                        "required": true,
                        "in": "path",
                        "description": "Asset group identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStockAssetBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateStockAssetForAssetGroupResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "DuplicateEntryError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "Finance API",
        "description": "Finance API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "scopes": {
                            "openid": "openid",
                            "profile": "profile",
                            "email": "email"
                        },
                        "authorizationUrl": "https://dev-tzti25ivp5zmox2s.us.auth0.com/authorize"
                    }
                },
                "in": "header",
                "name": "Authorization",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "PaginatedResponsePage": {
                "type": "object",
                "properties": {
                    "count": {
                        "type": "number",
                        "description": "The amount available in this response",
                        "example": 10
                    },
                    "offset": {
                        "type": "number",
                        "description": "The starting index of the paginated data",
                        "example": 0
                    },
                    "total": {
                        "type": "number",
                        "description": "The total number of items available",
                        "example": 10
                    }
                },
                "required": [
                    "count",
                    "offset",
                    "total"
                ]
            },
            "StockDataResponse": {
                "type": "object",
                "properties": {
                    "identity": {
                        "type": "string"
                    },
                    "symbol": {
                        "type": "string"
                    },
                    "exchange": {
                        "type": "string"
                    },
                    "assetKind": {
                        "type": "string",
                        "enum": [
                            "CS",
                            "ADRC",
                            "ADRP",
                            "ADRR",
                            "UNIT",
                            "RIGHT",
                            "PFD",
                            "FUND",
                            "SP",
                            "WARRANT",
                            "INDEX",
                            "ETF",
                            "ETN",
                            "OS",
                            "GDR",
                            "OTHER",
                            "NYRS",
                            "AGEN",
                            "EQLK",
                            "BOND",
                            "ADRW",
                            "BASKET",
                            "LT"
                        ]
                    }
                },
                "required": [
                    "identity",
                    "symbol",
                    "exchange",
                    "assetKind"
                ]
            },
            "GetSearchStockPaginatedResponse": {
                "type": "object",
                "properties": {
                    "page": {
                        "$ref": "#/components/schemas/PaginatedResponsePage"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StockDataResponse"
                        }
                    },
                    "isEmpty": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "page",
                    "data",
                    "isEmpty"
                ]
            },
            "GetSearchStockResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/GetSearchStockPaginatedResponse"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
            "HttpErrorResponse": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "integer",
                        "description": "The http code returned by the server"
                    },
                    "message": {
                        "type": "string",
                        "description": "A message describing the error that occurred"
                    }
                },
                "required": [
                    "statusCode",
                    "message"
                ]
            },
            "StockValueResponse": {
                "type": "object",
                "properties": {
                    "open": {
                        "type": "number"
                    },
                    "high": {
                        "type": "number"
                    },
                    "low": {
                        "type": "number"
                    },
                    "close": {
                        "type": "number"
                    },
                    "volume": {
                        "type": "integer"
                    }
                },
                "required": [
                    "open",
                    "high",
                    "low",
                    "close",
                    "volume"
                ]
            },
            "StockValuePaginatedResponse": {
                "type": "object",
                "properties": {
                    "page": {
                        "$ref": "#/components/schemas/PaginatedResponsePage"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StockValueResponse"
                        }
                    }
                },
                "required": [
                    "page",
                    "data"
                ]
            },
            "StockDataViewResponseData": {
                "type": "object",
                "properties": {
                    "stockData": {
                        "$ref": "#/components/schemas/StockDataResponse"
                    },
                    "yearlyValues": {
                        "$ref": "#/components/schemas/StockValuePaginatedResponse"
                    },
                    "hasValues": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "stockData",
                    "yearlyValues",
                    "hasValues"
                ]
            },
            "StockDataViewResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/StockDataViewResponseData"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
            "GetStockListResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/GetSearchStockPaginatedResponse"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
...
            "StockOrderResponse": {
                "type": "object",
                "properties": {
                    "amount": {
                        "type": "number"
                    },
                    "usdPrice": {
                        "type": "number"
                    }
                },
                "required": [
                    "amount",
                    "usdPrice"
                ]
            },
            "StockAssetResponse": {
                "type": "object",
                "properties": {
                    "identity": {
                        "type": "string"
                    },
                    "orders": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StockOrderResponse"
                        }
                    },
                    "stockData": {
                        "$ref": "#/components/schemas/StockDataResponse"
                    }
                },
                "required": [
                    "identity",
                    "orders",
                    "stockData"
                ]
            },
            "RealEstateAssetResponse": {
                "type": "object",
                "properties": {
                    "identity": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    }
                },
                "required": [
                    "identity",
                    "address"
                ]
            },
            "AssetResponse": {
                "type": "object",
                "properties": {
                    "identity": {
                        "type": "string"
                    },
                    "stockAsset": {
                        "nullable": true,
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/StockAssetResponse"
                            }
                        ]
                    },
                    "realEstateAsset": {
                        "nullable": true,
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/RealEstateAssetResponse"
                            }
                        ]
                    }
                },
                "required": [
                    "identity",
                    "stockAsset",
                    "realEstateAsset"
                ]
            },
...
            "StockOrderBody": {
                "type": "object",
                "properties": {
                    "price": {
                        "type": "number",
                        "minimum": 0
                    },
                    "amount": {
                        "type": "number",
                        "minimum": 0
                    }
                },
                "required": [
                    "price",
                    "amount"
                ]
            },
            "CreateStockAssetBody": {
                "type": "object",
                "properties": {
                    "stockDataIdentity": {
                        "type": "string",
                        "minLength": 5
                    },
                    "stockOrders": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StockOrderBody",
                            "minLength": 5
                        }
                    }
                },
                "required": [
                    "stockDataIdentity",
                    "stockOrders"
                ]
            },
            "CreateStockAssetForUserResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AssetResponse"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
...
            "CreateStockAssetForAssetGroupResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AssetResponse"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
...
        }
    }
}